fix(api): accept fractional Withings measure values and scale them exactly
WithingsMeasure.Value was an int, so a value with a fractional part failed
deserialization of the whole page and the sync reported Unknown. Read the
value as a decimal and apply the unit exponent with decimal arithmetic
instead of going through Math.Pow on doubles, so 79350 at unit -3 is exactly
79.35 and 79350.5 is 79.3505.

+++ apps/api/TrendWeight/Features/Providers/Withings/Models/WithingsMeasure.cs
@@ -8,10 +8,12 @@ namespace TrendWeight.Features.Providers.Withings.Models;
 public class WithingsMeasure
 {
     /// <summary>
-    /// The value of the measurement (needs to be converted using the unit)
+    /// The value of the measurement (needs to be converted using the unit).
+    /// Documented as an integer, but read as a decimal so a fractional value
+    /// does not fail the whole page.
     /// </summary>
     [JsonPropertyName("value")]
-    public int Value { get; set; }
+    public decimal Value { get; set; }
 
     /// <summary>
     /// The type of measurement:
